+++ op-service/retry/strategies_test.go
@@ -1,4 +1,4 @@
-package backoff
+package retry
 
 import (
 	"testing"
